Day 12 Part 2 Moon periods
I had the idea for this but thought it was crazy so I looked for a hint
on google and another guy used this approach too so I went ahead with
it. The strategy is work out the period of each axis and then find the
lowest common multiple of that.