Day 15 Part 2 Oxygen fill time
I realised the a-star algorithm scores the grid with a least path number
from the start position so I knew I could use it with a starting point
on the oxygen coordinate.

At first I thought I could re-use state from the a-star algorithm for
any unscored parts of the grid but it's hard because some of the state
is based on the target coordinate so I scrapped that idea and just did a
walk through all the un-scored coordinates. Since the walk started on an
outer boundary, it was pretty efficient anyway because it was scoring a
lot of nodes in-between the top edge and the oxygen.